% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grogger-package.R
\docType{package}
\name{grogger-package}
\alias{grogger}
\alias{grogger-package}
\title{grogger: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
