% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{swing}
\alias{swing}
\title{Build Grogger Blog}
\usage{
swing(
  index = here("index.Rmd"),
  posts = here("posts"),
  tags = TRUE,
  output_dir = here("docs")
)
}
\arguments{
\item{index}{Path to the \code{index.Rmd} file.}

\item{posts}{Path to the directory that contains the blog posts.}

\item{tags}{Create a page that collates posts according to tags?}

\item{output_dir}{Path to the directory where the output files will be
created.}
}
\description{
Build Grogger Blog
}
